refactor(kimi): separate session setup and typed native requests
