+++ backend/internal/integration/agents/kimi/server_commands.go
@@ -2,7 +2,6 @@ package kimi
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"strconv"
 	"strings"
@@ -26,12 +25,14 @@ func (r *serverRun) submit(ctx context.Context, p *serverTransport) (bool, error
 		prefix = strings.TrimSuffix(prompt, r.req.UserPrompt)
 	}
 
-	body := map[string]any{"disabled_tools": []string{}, "content": []any{map[string]any{"type": "text", "text": prompt}}}
+	disabledTools := []string{}
+	body := nativePrompt{DisabledTools: &disabledTools}
+	body.setText(prompt)
 	if !r.req.EnableBrowser {
-		body["disabled_tools"] = []string{"mcp__remote_browser__*"}
+		disabledTools = []string{"mcp__remote_browser__*"}
 	}
-	profile := func(config map[string]any) error {
-		return p.api(ctx, "POST", r.path()+"/profile", map[string]any{"agent_config": config}, nil)
+	profile := func(config nativeAgentConfig) error {
+		return p.api(ctx, "POST", r.path()+"/profile", nativeProfileUpdate{AgentConfig: &config}, nil)
 	}
 	done := func(message string, err error) (bool, error) {
 		if err == nil {
@@ -58,15 +59,16 @@ func (r *serverRun) submit(ctx context.Context, p *serverTransport) (bool, error
 		c.name, c.description, c.sideConversation = "Side question", question, true
 		r.childEvent(c, nil)
 		r.mainEnded = true
-		body["agent_id"] = side.AgentID
-		body["content"] = []any{map[string]any{"type": "text", "text": prefix + question}}
-		delete(body, "disabled_tools") // Keep the native side agent's tool policy.
+		body.AgentID = &side.AgentID
+		body.setText(prefix + question)
+		body.DisabledTools = nil // Keep the native side agent's tool policy.
 		return false, p.api(ctx, "POST", r.path()+"/prompts", body, nil)
 	case "/tower":
 		if args != "on" && args != "off" {
 			return true, fmt.Errorf("use /tower on or /tower off")
 		}
-		if err := profile(map[string]any{"tower_mode": args == "on"}); err != nil {
+		enabled := args == "on"
+		if err := profile(nativeAgentConfig{TowerMode: &enabled}); err != nil {
 			return true, err
 		}
 		if !hasRest {
@@ -77,22 +79,22 @@ func (r *serverRun) submit(ctx context.Context, p *serverTransport) (bool, error
 		if args != "" && args != "on" && args != "off" {
 			return true, fmt.Errorf("use /swarm on or /swarm off, followed by an optional prompt on the next line")
 		}
-		if err := profile(map[string]any{"swarm_mode": enabled}); err != nil {
+		if err := profile(nativeAgentConfig{SwarmMode: &enabled}); err != nil {
 			return true, err
 		}
 		if !hasRest {
 			return done("Kimi swarm mode "+map[bool]string{true: "enabled.", false: "disabled."}[enabled], nil)
 		}
 	case "/goal":
-		config := map[string]any{}
+		config := nativeAgentConfig{}
 		switch args {
 		case "pause", "resume", "cancel":
-			config["goal_control"] = args
+			config.GoalControl = args
 		default:
 			if args == "" {
 				return true, fmt.Errorf("use /goal <objective>, /goal pause, /goal resume, or /goal cancel")
 			}
-			config["goal_objective"] = args
+			config.GoalObjective = args
 		}
 		if err := profile(config); err != nil {
 			return true, err
@@ -107,10 +109,10 @@ func (r *serverRun) submit(ctx context.Context, p *serverTransport) (bool, error
 		if args == "" || !hasRest || strings.TrimSpace(rest) == "" {
 			return true, fmt.Errorf("use /agent <profile> followed by a prompt on the next line")
 		}
-		body["profile"] = args
-		body["model"] = r.req.Model
+		body.Profile = args
+		body.Model = &r.req.Model
 		if r.thinking != "" {
-			body["thinking"] = r.thinking
+			body.Thinking = r.thinking
 		}
 	case "/compact":
 		r.compacting = true
@@ -137,46 +139,14 @@ func (r *serverRun) submit(ctx context.Context, p *serverTransport) (bool, error
 			if name == "" {
 				return true, fmt.Errorf("missing Kimi skill name")
 			}
-			body["skills"] = []any{map[string]any{"name": name, "args": strings.TrimSpace(args + "\n" + rest)}}
-			body["content"] = []any{map[string]any{"type": "text", "text": prefix + "Use the selected skill."}}
+			body.Skills = []nativeSkillActivation{{Name: name, Args: strings.TrimSpace(args + "\n" + rest)}}
+			body.setText(prefix + "Use the selected skill.")
 			return false, p.api(ctx, "POST", r.path()+"/prompts", body, nil)
 		}
 		return false, p.api(ctx, "POST", r.path()+"/prompts", body, nil)
 	}
-	body["content"] = []any{map[string]any{"type": "text", "text": prefix + rest}}
+	body.setText(prefix + rest)
 	return false, p.api(ctx, "POST", r.path()+"/prompts", body, nil)
-}
-
-const remoteBrowserName = "remote_browser"
-
-func (r *serverRun) prepareBrowser(ctx context.Context, p *serverTransport) error {
-	if !r.req.EnableBrowser {
-		return nil
-	}
-	desired := map[string]any{"transport": "stdio", "command": "npx", "args": []string{"@playwright/mcp", "--cdp-endpoint", "http://127.0.0.1:9222", "--caps=vision"}}
-	var existing struct {
-		Config struct {
-			Command string   `json:"command"`
-			Args    []string `json:"args"`
-		} `json:"config"`
-	}
-	err := p.api(ctx, "GET", "/api/v2/mcp/servers/"+remoteBrowserName, nil, &existing)
-	if err == nil {
-		args, _ := json.Marshal(existing.Config.Args)
-		want, _ := json.Marshal(desired["args"])
-		if existing.Config.Command != "npx" || string(args) != string(want) {
-			return fmt.Errorf("Kimi MCP name %q is already configured differently", remoteBrowserName)
-		}
-		return nil
-	}
-	apiErr, ok := err.(*serverError)
-	if !ok || apiErr.Code != 40408 {
-		return err
-	}
-	// The dedicated entry is merged by Kimi's management API, which preserves
-	// unrelated user/project/plugin entries and refuses read-only collisions.
-	desired["name"] = remoteBrowserName
-	return p.api(ctx, "POST", "/api/v2/mcp/servers", desired, nil)
 }
 
 func (r *serverRun) userCommand() string {
